+++ libnavui-maps/src/main/java/com/mapbox/navigation/ui/maps/route/line/api/MapboxRouteLineApi.kt
@@ -28,10 +28,13 @@ import com.mapbox.navigation.core.routealternatives.AlternativeRouteMetadata
 import com.mapbox.navigation.core.trip.session.RouteProgressObserver
 import com.mapbox.navigation.ui.base.util.MapboxNavigationConsumer
 import com.mapbox.navigation.ui.maps.internal.route.line.MapboxRouteLineUtils
+import com.mapbox.navigation.ui.maps.internal.route.line.MapboxRouteLineUtils.distinctGranularDistancesProvider
 import com.mapbox.navigation.ui.maps.internal.route.line.MapboxRouteLineUtils.getMatchingColors
+import com.mapbox.navigation.ui.maps.internal.route.line.MapboxRouteLineUtils.granularDistancesProvider
 import com.mapbox.navigation.ui.maps.internal.route.line.MapboxRouteLineUtils.layerGroup1SourceLayerIds
 import com.mapbox.navigation.ui.maps.internal.route.line.MapboxRouteLineUtils.layerGroup2SourceLayerIds
 import com.mapbox.navigation.ui.maps.internal.route.line.MapboxRouteLineUtils.layerGroup3SourceLayerIds
+import com.mapbox.navigation.ui.maps.internal.route.line.MapboxRouteLineUtils.routePointsProvider
 import com.mapbox.navigation.ui.maps.route.RouteLayerConstants
 import com.mapbox.navigation.ui.maps.route.line.model.ClosestRouteValue
 import com.mapbox.navigation.ui.maps.route.line.model.ExtractedRouteData
@@ -67,6 +70,7 @@ import kotlinx.coroutines.cancelChildren
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
+import kotlinx.coroutines.withContext
 import java.util.concurrent.CopyOnWriteArrayList
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
@@ -453,39 +457,45 @@ class MapboxRouteLineApi(
             RouteLayerConstants.SOFT_GRADIENT_STOP_GAP_METERS / route.distance()
         } ?: .00000000001 // an arbitrarily small value so Expression values are in ascending order
 
-        val routeLineExpressionProviders =
-            if (routeLineOptions.styleInactiveRouteLegsIndependently) {
-                val workingRouteLineExpressionData =
-                    alternativelyStyleSegmentsNotInLeg(activeLegIndex, routeLineExpressionData)
+        val routeLineExpressionProviders = ifNonNull(primaryRoute) { route ->
+            ifNonNull(granularDistancesProvider(route)) { granularDistances ->
+                if (routeLineOptions.styleInactiveRouteLegsIndependently) {
+                    val workingRouteLineExpressionData =
+                        alternativelyStyleSegmentsNotInLeg(activeLegIndex, routeLineExpressionData)
 
-                val restrictedExpressionData: List<ExtractedRouteData>? =
-                    ifNonNull(primaryRoute?.directionsRoute) { route ->
+                    val restrictedExpressionData: List<ExtractedRouteData>? =
+
                         if (routeLineOptions.displayRestrictedRoadSections &&
                             routeHasRestrictions
                         ) {
                             MapboxRouteLineUtils.extractRouteData(
-                                route,
+                                route.directionsRoute,
                                 MapboxRouteLineUtils.getTrafficCongestionAnnotationProvider(
-                                    route,
+                                    route.directionsRoute,
                                     routeLineOptions.resourceProvider.routeLineColorResources
                                 )
                             )
                         } else {
                             null
                         }
-                    }
-                routeLineOptions.vanishingRouteLine?.getTraveledRouteLineExpressions(
-                    point,
-                    workingRouteLineExpressionData,
-                    restrictedExpressionData,
-                    routeLineOptions.resourceProvider,
-                    activeLegIndex,
-                    stopGap,
-                    routeLineOptions.displaySoftGradientForTraffic,
-                )
-            } else {
-                routeLineOptions.vanishingRouteLine?.getTraveledRouteLineExpressions(point)
+                    routeLineOptions.vanishingRouteLine?.getTraveledRouteLineExpressions(
+                        point,
+                        granularDistances,
+                        workingRouteLineExpressionData,
+                        restrictedExpressionData,
+                        routeLineOptions.resourceProvider,
+                        activeLegIndex,
+                        stopGap,
+                        routeLineOptions.displaySoftGradientForTraffic,
+                    )
+                } else {
+                    routeLineOptions.vanishingRouteLine?.getTraveledRouteLineExpressions(
+                        point,
+                        granularDistances
+                    )
+                }
             }
+        }
 
         lastPointUpdateTimeNano = System.nanoTime()
         return when (routeLineExpressionProviders) {
@@ -546,7 +556,6 @@ class MapboxRouteLineApi(
     ) {
         jobControl.scope.launch(Dispatchers.Main) {
             mutex.withLock {
-                routeLineOptions.vanishingRouteLine?.clear()
                 routeLineOptions.vanishingRouteLine?.vanishPointOffset = 0.0
                 activeLegIndex = INVALID_ACTIVE_LEG_INDEX
                 routes.clear()
@@ -954,7 +963,6 @@ class MapboxRouteLineApi(
      * Cancels any/all background tasks that may be running.
      */
     fun cancel() {
-        routeLineOptions.vanishingRouteLine?.cancel()
         jobControl.job.cancelChildren()
     }
 
@@ -1076,7 +1084,7 @@ class MapboxRouteLineApi(
         ifNonNull(
             routeProgress.currentLegProgress,
             routeProgress.currentLegProgress?.currentStepProgress,
-            routeLineOptions.vanishingRouteLine?.primaryRoutePoints
+            routePointsProvider(routeProgress.navigationRoute)
         ) { currentLegProgress, currentStepProgress, completeRoutePoints ->
             var allRemainingPoints = 0
 
@@ -1149,10 +1157,12 @@ class MapboxRouteLineApi(
                 LOG_CATEGORY
             )
         }
+        val distinctAlternativeRouteMetadata = alternativeRoutesMetadata.filter { metadata ->
+            distinctNewRoutes.find { it.id == metadata.navigationRoute.id } != null
+        }
 
         ifNonNull(distinctNewRoutes.firstOrNull()) { primaryRouteCandidate ->
             if (!primaryRouteCandidate.directionsRoute.isSameRoute(primaryRoute?.directionsRoute)) {
-                routeLineOptions.vanishingRouteLine?.clear()
                 routeLineOptions.vanishingRouteLine?.vanishPointOffset = 0.0
             }
         }
@@ -1161,10 +1171,36 @@ class MapboxRouteLineApi(
         routes.addAll(distinctNewRoutes)
         primaryRoute = distinctNewRoutes.firstOrNull()
         MapboxRouteLineUtils.trimRouteDataCacheToSize(size = distinctNewRoutes.size)
-        alternativesDeviationOffset =
-            MapboxRouteLineUtils.getAlternativeRoutesDeviationOffsets(alternativeRoutesMetadata)
+
+        preWarmRouteCaches(
+            distinctNewRoutes,
+            vanishingRouteLineEnabled = routeLineOptions.vanishingRouteLine != null,
+            alternativeRouteMetadataAvailable = distinctAlternativeRouteMetadata.isNotEmpty()
+        )
+
+        alternativesDeviationOffset = distinctAlternativeRouteMetadata.associate {
+            it.navigationRoute.id to MapboxRouteLineUtils.getAlternativeRouteDeviationOffsets(it)
+        }
 
         return buildDrawRoutesState(featureDataProvider)
+    }
+
+    private suspend fun preWarmRouteCaches(
+        routes: List<NavigationRoute>,
+        vanishingRouteLineEnabled: Boolean,
+        alternativeRouteMetadataAvailable: Boolean
+    ) {
+        if (routes.isEmpty()) return
+        withContext(jobControl.scope.coroutineContext) {
+            if (vanishingRouteLineEnabled) {
+                granularDistancesProvider(routes.first())
+            }
+            if (alternativeRouteMetadataAvailable) {
+                routes.drop(1).forEach {
+                    distinctGranularDistancesProvider(it)
+                }
+            }
+        }
     }
 
     private suspend fun buildDrawRoutesState(
@@ -1463,20 +1499,6 @@ class MapboxRouteLineApi(
                     } ?: listOf()
                 }
                 routeLineExpressionData = segmentsDef.await()
-            }
-        }
-
-        // This call is resource intensive so it needs to come last so that
-        // it doesn't consume resources used by the calculations above. The results
-        // of this call aren't necessary to return to the caller but the calculations above are.
-        // Putting this call above will delay the caller receiving the result. There is a check
-        // for primaryRouteLineGranularDistances being null because this value is cleared only
-        // if the primary route has changed via setRoutes. If there's no change in the primary
-        // route there's no need to re-init the vanishing route line which is an expensive
-        // operation.
-        if (routeLineOptions.vanishingRouteLine?.primaryRouteLineGranularDistances == null) {
-            partitionedRoutes.first.firstOrNull()?.let {
-                routeLineOptions.vanishingRouteLine?.initWithRoute(it.route)
             }
         }